restructure to multi-module
