+++ bitstream-reader/src/main/java/org/danskells/bitstream/BitContainer.java
@@ -1,6 +1,6 @@
 package org.danskells.bitstream;
 
-public interface ReadableBitContainer {
+public interface BitContainer {
 
     StreamNode getStreamNode(long startInclusive, long endExclusive);
