+++ bitstream-reader/src/main/java/org/danskells/bitstream/BitContainerStream.java
