+++ bitstream-reader/src/main/java/org/danskells/bitstream/Block.java
@@ -1,0 +1,10 @@
+package org.danskells.bitstream;
+/**
+ * A Block represents a segment of a BitContainer.
+ * It's a structure that holds a collection of bits within a specific range.
+ * It provides a StreamNode to iterate over the bits set within that block.
+ */
+public interface Block {
+
+    StreamNode getStreamNode();
+}
