+++ bitstream-reader/src/main/java/org/danskells/bitstream/StreamNode.java
