+++ bitstream-reader/src/main/java/org/danskells/bitstream/streamnode/AndStreamNode.java
