+++ bitstream-reader/src/main/java/org/danskells/bitstream/streamnode/OrStreamNode.java
