+++ bitstream-reader/src/test/java/org/danskells/bitstream/block/LongArrayBlockTest.java
@@ -1,4 +1,4 @@
-package org.danskells.bitstream;
+package org.danskells.bitstream.block;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
