+++ bitstream-reader/src/test/java/org/danskells/bitstream/container/SimpleReadableBitContainerTest.java
@@ -1,6 +1,9 @@
-package org.danskells.bitstream;
+package org.danskells.bitstream.container;
 
-import org.junit.Test;
+import org.danskells.bitstream.Block;
+import org.danskells.bitstream.block.BitSetBlockTest;
+import org.danskells.bitstream.block.BitmapBlock;
+import org.danskells.bitstream.container.SimpleReadableBitContainer;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
