--- src/main/java/org/danskells/bitstream/Block.java
@@ -1,6 +1,0 @@
-package org.danskells.bitstream;
-
-public interface Block {
-
-    public StreamNode getStreamNode();
-}
